Improve Ex613.

+++ src/ex613/Point.java
@@ -1,10 +1,10 @@
 package ex613;
 
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
